+++ src/context/context.tsx
@@ -1,3 +1,4 @@
 import React from 'react';
 
 export const Context = React.createContext({} as { (url?: string): void });
+export const SearchContext = React.createContext({});
